% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kliqfind.R
\name{check_int}
\alias{check_int}
\title{Check that input is of type integer; if not, assert that it only contains
whole numbers and convert to integer type}
\usage{
check_int(input)
}
\description{
Check that input is of type integer; if not, assert that it only contains
whole numbers and convert to integer type
}
