% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doskliq.R
\name{doskliq_exe_path}
\alias{doskliq_exe_path}
\title{Get path to doskliq executable}
\usage{
doskliq_exe_path()
}
\value{
A string representing the path to the doskliq executable
}
\description{
Get path to doskliq executable
}
