% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doskliq.R
\name{doskliq_run}
\alias{doskliq_run}
\title{Run doskliq executable}
\usage{
doskliq_run(input_path, output_path)
}
\arguments{
\item{input_path}{The path to read the input files from}

\item{output_path}{The path to write the output files to}
}
\description{
Run doskliq executable
}
