% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kliqfind.R
\name{get_plc}
\alias{get_plc}
\title{Calculate the subgroup memberships}
\usage{
get_plc(res)
}
\arguments{
\item{res}{A result list from the KliqueFinder Fortran code}
}
\value{
A data.frame with actor IDs, subgroup membership, and paired actor
}
\description{
Calculate the subgroup memberships
}
