% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkliq.R
\name{kf_locate_input}
\alias{kf_locate_input}
\title{Get the path to a kliqfindr input file}
\usage{
kf_locate_input(input_file)
}
\arguments{
\item{input_file}{The path to an input file or name of file in extdata}
}
\value{
The path to the input file
}
\description{
Get the path to a kliqfindr input file
}
