% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kliqfind.R
\name{klique_find}
\alias{klique_find}
\title{Run the KliqueFinder algorithm}
\usage{
klique_find(datamat, numact, maxiter = as.integer(999),
  attachi = as.integer(1), output_path = ".")
}
\arguments{
\item{datamat}{The matrix of contact frequency}

\item{numact}{The number of actors}

\item{maxiter}{The maximum number of iterations to go through before
stopping on account of non-convergence}

\item{attachi}{Flag if set to 1 isolates are to be added even if it results
in a decrease in theta1}

\item{output_path}{The path to write the output files to}
}
\value{
The result list from the KliqueFinder Fortran code
}
\description{
Run the KliqueFinder algorithm
}
