% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkliq.R
\name{read_place}
\alias{read_place}
\title{Read a place file}
\usage{
read_place(place_file)
}
\arguments{
\item{place_file}{The path to the place file}
}
\value{
A data frame with each actor's subgroup assignment
}
\description{
Read a place file
}
