% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkliq.R
\name{smacof1b_exe_path}
\alias{smacof1b_exe_path}
\title{Get path to smacof1b executable}
\usage{
smacof1b_exe_path()
}
\value{
A string representing the path to the smacof1b executable
}
\description{
Get path to smacof1b executable
}
