% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkliq.R
\name{winkliq_exe_path}
\alias{winkliq_exe_path}
\title{Get path to winkliq executable}
\usage{
winkliq_exe_path()
}
\value{
A string representing the path to the winkliq executable
}
\description{
Get path to winkliq executable
}
