% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkliq.R
\name{winkliq_run}
\alias{winkliq_run}
\title{Run winkliq executable}
\usage{
winkliq_run(input_file, printo_file = "basic.printo",
  param_file = "basic.param", working_dir = NA, get_place = TRUE,
  overwrite = TRUE)
}
\arguments{
\item{input_file}{The path to the input datafile}

\item{printo_file}{The path to the printo config file}

\item{param_file}{The path to the kliqfind.par config file}

\item{working_dir}{The path to the working directory}

\item{get_place}{Whether to read the place file}

\item{overwrite}{Whether to overwrite input files in working directory}
}
\value{
A list containing the error code from system2 (0 for success),
        the output directory, and optionally the place file output read in
        as a data frame
}
\description{
Run winkliq executable
}
\examples{
winkliq_run('stanne.list', 'basic.printo', 'basic.param', get_place=FALSE)
}
